Made exporting a graphi as a tikz input a free function.
